update libft, reorganize files, add optimization flags
