+++ inc/structs.h
@@ -6,7 +6,7 @@
 /*   By: maolivei <maolivei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/17 21:06:43 by ralves-b          #+#    #+#             */
-/*   Updated: 2022/11/02 15:02:01 by maolivei         ###   ########.fr       */
+/*   Updated: 2022/11/04 10:21:16 by maolivei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
