+++ libs/libft/Makefile
@@ -33,14 +33,16 @@ SOURCE_FILES		+= ft_putstr_fd.c ft_putendl_fd.c ft_putnbr_fd.c
 SOURCE_FILES		+= ft_putchar_fd.c
 
 # stdlib
-SOURCE_FILES		+= ft_atoi.c ft_atoll.c ft_calloc.c
+SOURCE_FILES		+= ft_atof.c ft_atoi.c ft_atoll.c ft_calloc.c
 
 # custom
 SOURCE_FILES		+= ft_gnl.c ft_gnl_multifd.c ft_memfree.c ft_free_matrix.c
 SOURCE_FILES		+= ft_skip_chars.c ft_split.c ft_striteri.c ft_strmapi.c
 SOURCE_FILES		+= ft_strjoin.c ft_strjoin_free.c ft_strjoin_free_null.c
 SOURCE_FILES		+= ft_strtrim.c ft_strtrim_free.c ft_strtrim_free_null.c
-SOURCE_FILES		+= ft_substr.c ft_stredit.c
+SOURCE_FILES		+= ft_substr.c ft_stredit.c ft_splitsize.c ft_isfloat.c
+SOURCE_FILES		+= ft_isnumber.c ft_isinrange_f.c ft_isinrange_i.c
+SOURCE_FILES		+= ft_swap_set.c
 
 # non_standard
 SOURCE_FILES		+= ft_itoa_base.c ft_itoa.c ft_strlcat.c ft_strlcpy.c
@@ -67,7 +69,7 @@ CFLAGS				:= -Wall -Wextra -Werror
 IFLAGS				:= -I $(HEADER_PATH)
 REMOVE				:= rm -rf
 
-ifdef C_DEBUG
+ifdef DEBUG
 	CFLAGS += -g3
 endif
