+++ libs/libft/sources/custom/ft_isinrange_i.c
@@ -1,26 +1,18 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   ft_str_swap_set_chr.c                              :+:      :+:    :+:   */
+/*   ft_isinrange_i.c                                   :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ralves-b <ralves-b@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: maolivei <maolivei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/18 16:29:43 by ralves-b          #+#    #+#             */
-/*   Updated: 2022/10/18 16:41:30 by ralves-b         ###   ########.fr       */
+/*   Created: 2022/11/04 10:32:25 by maolivei          #+#    #+#             */
+/*   Updated: 2022/11/04 10:52:42 by maolivei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-#include <minirt.h>
+#include "libft.h"
 
-void	ft_str_swap_set_chr(char *str, char *old, char new)
+t_bool	ft_isinrange_i(int value, int min, int max)
 {
-	int	i;
-
-	i = 0;
-	while (str[i])
-	{
-		if (ft_strchr(old, str[i]))
-			str[i] = new;
-		i++;
-	}
+	return (value >= min && value <= max);
 }
