+++ src/parser/parser.c
@@ -3,145 +3,92 @@
 /*                                                        :::      ::::::::   */
 /*   parser.c                                           :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: ralves-b <ralves-b@student.42sp.org.br>    +#+  +:+       +#+        */
+/*   By: maolivei <maolivei@student.42sp.org.br>    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2022/10/17 19:29:07 by ralves-b          #+#    #+#             */
-/*   Updated: 2022/11/01 11:22:39 by ralves-b         ###   ########.fr       */
+/*   Created: 2022/11/01 20:51:52 by maolivei          #+#    #+#             */
+/*   Updated: 2022/11/04 10:49:51 by maolivei         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include <minirt.h>
 
-void	cavalinho_cai_matando_geral(t_world *w, t_cam *camera, t_scene *scene)
-{
-	t_list	*temp;
+#define ERR_OPEN_FAIL "Could not open .rt file"
+#define ERR_BAD_ID "Invalid identifier in .rt file."
 
-	destroy_world(w);
-	destroy_camera(camera);
-	free(scene->light->rgb);
-	free(scene->light);
-	free(scene->camera);
-	free(scene->ambience->rgb);
-	free(scene->ambience);
-	while (scene->objects)
+static t_delegator	*get_parse_function(t_parse_id id)
+{
+	static t_delegator	*function_table[PARSE_ID_MAX + 1];
+
+	if (function_table[PARSE_ID_MIN])
+		return (function_table[id]);
+	function_table[PARSE_AMBIENT] = parse_ambient;
+	function_table[PARSE_CAMERA] = parse_camera;
+	function_table[PARSE_LIGHT] = parse_light;
+	function_table[PARSE_SPHERE] = parse_sphere;
+	function_table[PARSE_PLANE] = parse_plane;
+	function_table[PARSE_CYLINDER] = parse_cylinder;
+	return (function_table[id]);
+}
+
+static t_parse_id	get_token_id(char *token)
+{
+	if (ft_strcmp(AMBIENT, token) == 0)
+		return (PARSE_AMBIENT);
+	if (ft_strcmp(CAMERA, token) == 0)
+		return (PARSE_CAMERA);
+	if (ft_strcmp(LIGHT, token) == 0)
+		return (PARSE_LIGHT);
+	if (ft_strcmp(SPHERE, token) == 0)
+		return (PARSE_SPHERE);
+	if (ft_strcmp(CYLINDER, token) == 0)
+		return (PARSE_CYLINDER);
+	if (ft_strcmp(PLANE, token) == 0)
+		return (PARSE_PLANE);
+	return (NONE);
+}
+
+static int	parse_rt_file(char *line, t_rt_scene *s)
+{
+	int			status;
+	char		**tokens;
+	t_delegator	*parse_function;
+
+	ft_swap_set(line, BLANKSPACES, ' ');
+	tokens = ft_split(line, ' ');
+	if (!tokens)
+		return (0);
+	parse_function = get_parse_function(get_token_id(tokens[0]));
+	if (!parse_function)
+		return (ft_free_matrix((void *)&tokens), error(ERR_BAD_ID));
+	status = parse_function(tokens, s);
+	ft_free_matrix((void *)&tokens);
+	return (status);
+}
+
+int	read_rt_file(char *filename, t_rt_scene *s)
+{
+	int		file_descriptor;
+	char	*current_line;
+
+	if (check_file_extension(filename) != 0)
+		return (-1);
+	file_descriptor = open(filename, O_RDONLY);
+	if (file_descriptor < 0)
+		return (perror(ERR_OPEN_FAIL), -1);
+	while (TRUE)
 	{
-		temp = scene->objects->next;
-		free(scene->objects);
-		scene->objects = temp;
-	}
-	free(scene);
-}
-
-int	end_program(t_canvas *mlx)
-{
-	destroy_canvas(mlx);
-	exit(0);
-	return (0);
-}
-
-void	generate_objects(t_world *world, t_scene *scene)
-{
-	t_list		*lst_aux;
-	t_object	*obj_aux;
-
-	lst_aux = scene->objects;
-	while (lst_aux)
-	{
-		obj_aux = (t_object *)lst_aux->content;
-		ft_lstadd_front(&world->objects, ft_lstnew(obj_aux));
-		lst_aux = lst_aux->next;
-	}
-}
-
-void	generate_world(t_scene *s)
-{
-	t_world		*world;
-	t_light_pnt	*lp;
-	t_cam		*camera;
-	t_canvas	*canvas;
-
-	lp = create_light_point(
-			create_point(s->light->x, s->light->y, s->light->z), \
-		scalar_multiply_color(s->light->rgb, s->light->brightness));
-	world = create_world();
-	world->light_point = ft_lstnew(lp);
-	generate_objects(world, s);
-	camera = create_camera(1000, 500, radians(s->camera->fov));
-	set_camera_transformation(camera, view_transform(
-			&(t_point){s->camera->view_x, s->camera->view_y, \
-			s->camera->view_z, 1},
-			&(t_point){s->camera->x_3d, s->camera->y_3d, s->camera->z_3d, 1},
-			&(t_vector){0, 1, 0, 0}));
-	canvas = render(camera, world);
-	canvas->win = mlx_new_window(
-			canvas->mlx, camera->h_size, camera->v_size, "cavalinho assassino");
-	mlx_put_image_to_window(canvas->mlx, canvas->win, canvas->image, 0, 0);
-	cavalinho_cai_matando_geral(world, camera, scene);
-	mlx_hook(canvas->win, 17, (1L << 2), end_program, canvas);
-	mlx_loop(canvas->mlx);
-}
-
-int	sub_main(int argc, char **argv)
-{
-	t_scene	*scene;
-	int		fd;
-
-	if (errors(argc, argv, &fd))
-		return (EXIT_FAILURE);
-	scene = (t_scene *)ft_calloc(1, sizeof(t_scene));
-	if (parser_1(fd, scene))
-		return (EXIT_FAILURE);
-	generate_world(scene);
-	return (EXIT_SUCCESS);
-}
-
-int	parse_line(char **line_splited, t_scene *scene)
-{
-	int	errors;
-
-	errors = 0;
-	if (!ft_strcmp(line_splited[0], LIGHT) && !scene->light)
-		errors = check_light(line_splited, &scene->light);
-	else if (!ft_strcmp(line_splited[0], AMBIENT) && !scene->ambience)
-		errors = check_ambient(line_splited, &scene->ambience);
-	else if (!ft_strcmp(line_splited[0], CAMERA) && !scene->camera)
-		errors = check_camera(line_splited, &scene->camera);
-	else if (!ft_strcmp(line_splited[0], SPHERE))
-		errors = check_sphere(line_splited, &scene->objects);
-	else if (!ft_strcmp(line_splited[0], CYLINDER))
-		errors = check_cylinder(line_splited, &scene->objects);
-	else if (!ft_strcmp(line_splited[0], PLANE))
-		errors = check_plane(line_splited, &scene->objects);
-	else if (line_splited[0][0] != '#')
-		errors = 1;
-	if (!errors)
-		return (EXIT_SUCCESS);
-	return (EXIT_FAILURE);
-}
-
-int	parser_1(int fd, t_scene *scene)
-{
-	char		*line;
-	char		**line_splited;
-	int			errors;
-
-	errors = 0;
-	line = ft_gnl(fd, FALSE);
-	while (line)
-	{
-		line = ft_strtrim_free(line, " \f\n\r\t\v");
-		ft_str_swap_set_chr(line, "\f\r\t\v", ' ');
-		line_splited = ft_split(line, ' ');
-		free(line);
-		if (line_splited)
+		current_line = ft_gnl(file_descriptor, FALSE);
+		if (!current_line)
+			break ;
+		if (!*current_line || *current_line == '#')
 		{
-			if (parse_line(line_splited, scene))
-				errors++;
-			ft_free_matrix((void *)&line_splited);
+			free(current_line);
+			continue ;
 		}
-		line = ft_gnl(fd, FALSE);
+		if (parse_rt_file(current_line, s) != 0)
+			return (free(current_line), close(file_descriptor), -1);
+		free(current_line);
 	}
-	if (errors || !scene->camera || !scene->ambience || !scene->light)
-		return (free_scene(scene), EXIT_FAILURE);
-	return (EXIT_SUCCESS);
+	close(file_descriptor);
+	return (check_scene_elements(s));
 }
