+++ src/parser/scene_to_world.c
